requirements.txt
discord
requests
python-dotenv
datetime
pymongo
pytest